fix broken test

+++ integration/test_collection.py
@@ -1356,7 +1356,7 @@ def test_return_properties_with_typed_dict(client: weaviate.Client, which_case: 
 
         objects = collection.query.get_flat(return_properties=DataModel)
         assert len(objects) == 1
-        assert objects[0].properties == {"int": 1}
+        assert objects[0].properties == {"int_": 1}
     elif which_case == 1:
 
         class DataModel(TypedDict):
